+++ app/classes/observation_report/dwca.rb
@@ -2,12 +2,12 @@
 
 module ObservationReport
   # Darwin Core Archive format.
-  class Darwin < ObservationReport::ZipReport
+  class Dwca < ObservationReport::ZipReport
     attr_accessor :csv
 
     def initialize(args)
       super(args)
-      self.csv = DarwinCSV.new(args)
+      self.csv = Darwin::Observations.new(args)
     end
 
     def filename
